+++ ArcPicker/ArcPickerBehaviorDecorator.js
@@ -1,8 +1,12 @@
 import {forward} from '@enact/core/handle';
+import {is} from '@enact/core/keymap';
 import hoc from '@enact/core/hoc';
 import {validateRangeOnce} from '@enact/ui/internal/validators';
 import PropTypes from 'prop-types';
 import React from 'react';
+
+const isDown = is('down');
+const isUp = is('up');
 
 const validateRange = validateRangeOnce((props) => props, {'component': 'ArcPickerBehaviorDecorator'});
 
@@ -27,6 +31,14 @@ const ArcPickerBehaviorDecorator = hoc((config, Wrapped) => {
 			 * @public
 			 */
 			children: PropTypes.node.isRequired,
+
+			/**
+			 * Whether or not the component is in a disabled state.
+			 *
+			 * @type {Boolean}
+			 * @public
+			 */
+			disabled: PropTypes.bool,
 
 			/**
 			 * The maximum value of ArcPicker.
@@ -91,6 +103,21 @@ const ArcPickerBehaviorDecorator = hoc((config, Wrapped) => {
 			this.setState({isFocused: true});
 		};
 
+		handleKeyDown = (ev, props) => {
+			const {children, disabled, value} = this.props;
+			const index = children.findIndex(child => child === value);
+
+			forward('onKeyDown', ev, props);
+
+			if (!disabled) {
+				if (isDown(ev.keyCode)) {
+					this.handleClick(children[Math.max(index - 1, 0)])(ev);
+				} else if (isUp(ev.keyCode)) {
+					this.handleClick(children[Math.min(index + 1, children.length - 1)])(ev);
+				}
+			}
+		};
+
 		render () {
 			const {children, max, min, value: valueProp, ...rest} = this.props;
 			const value = ((valueProp || valueProp === 0) ? valueProp : children[0]);
@@ -110,6 +137,7 @@ const ArcPickerBehaviorDecorator = hoc((config, Wrapped) => {
 					onBlur={this.handleBlur}
 					onClick={this.handleClick}
 					onFocus={this.handleFocus}
+					onKeyDown={this.handleKeyDown}
 					value={value}
 				>
 					{children}
