+++ ArcSlider/ArcSliderBehaviorDecorator.js
@@ -1,4 +1,5 @@
 import {forward} from '@enact/core/handle';
+import {is} from '@enact/core/keymap';
 import hoc from '@enact/core/hoc';
 import {validateRangeOnce, validateSteppedOnce} from '@enact/ui/internal/validators';
 import PropTypes from 'prop-types';
@@ -7,6 +8,9 @@ import React from 'react';
 import {positionToAngle} from '../Arc/utils';
 
 import {angleToValue} from './utils';
+
+const isDown = is('down');
+const isUp = is('up');
 
 const validateValueRange = validateRangeOnce((props) => props, {'component': 'ArcSliderBehaviorDecorator'});
 const validateAngleRange = validateRangeOnce((props) => props, {'component': 'ArcSliderBehaviorDecorator', minName: 'startAngle', maxName: 'endAngle'});
@@ -26,6 +30,14 @@ const ArcSliderBehaviorDecorator = hoc((config, Wrapped) => {
 		static displayName = 'ArcSliderBehaviorDecorator';
 
 		static propTypes = /** @lends agate/ArcSlider.ArcSliderBehaviorDecorator.prototype */ {
+			/**
+			 * Whether or not the component is in a disabled state.
+			 *
+			 * @type {Boolean}
+			 * @public
+			 */
+			disabled: PropTypes.bool,
+
 			/**
 			 * The end angle(in degrees) of the arc slider.
 			 *
@@ -136,7 +148,10 @@ const ArcSliderBehaviorDecorator = hoc((config, Wrapped) => {
 		handleDown = ({clientX, clientY}) => {
 			const params = {x: clientX, y: clientY};
 			forward('onDown', params, this.props);
-			this.emitChangeForPosition(params);
+
+			if (clientX != null && clientY != null) {
+				this.emitChangeForPosition(params);
+			}
 		};
 
 		handleDragStart = (ev) => {
@@ -206,6 +221,21 @@ const ArcSliderBehaviorDecorator = hoc((config, Wrapped) => {
 			this.setState({isFocused: true});
 		};
 
+		handleKeyDown = (ev) => {
+			const {disabled, max, min, step} = this.props;
+			const {value} = this.state;
+
+			forward('onKeyDown', ev, this.props);
+
+			if (!disabled) {
+				if (isDown(ev.keyCode)) {
+					this.handleChange(ev, Math.max(value - step, min));
+				} else if (isUp(ev.keyCode)) {
+					this.handleChange(ev, Math.min(value + step, max));
+				}
+			}
+		};
+
 		render () {
 			if (__DEV__) {
 				const {endAngle, max, min, startAngle, step} = this.props;
@@ -229,6 +259,7 @@ const ArcSliderBehaviorDecorator = hoc((config, Wrapped) => {
 					onDrag={this.handleDrag}
 					onDragStart={this.handleDragStart}
 					onFocus={this.handleFocus}
+					onKeyDown={this.handleKeyDown}
 					value={this.state.value}
 				/>
 			);
